+++ packages/sdk/package.json
@@ -24,7 +24,7 @@
     "clean": "rimraf dist dist-* temp types *.tgz *.log",
     "format": "prettier --write --config ./.prettierrc.json --ignore-path ./.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"samples/**/*.ts\" \"*.{js,json}\"",
     "integration-test:browser-case": "karma start --single-run --glob=./dist-test/index.integration.browser.js",
-    "integration-test:browser": "concurrently --kill-others \"node test/integration/scripts/StartSimpleAuth.js\" \"npm run integration-test:browser-case\" || true",
+    "integration-test:browser": "node test/integration/scripts/DownloadSimpleAuth.js && concurrently --kill-others \"node test/integration/scripts/StartSimpleAuth.js\" \"npm run integration-test:browser-case\" || true",
     "integration-test:node": "mocha -r test/mocha.env.ts -r ts-node/register \"./test/integration/node/*.spec.ts\" --timeout 100000 --exit --reporter ./tools/mocha-multi-reporter.js",
     "test:integration": "npm run integration-test:node && npm run integration-test:browser",
     "lint:staged": "lint-staged",
